+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# v0.30.0
+
+## Features
+* feat(lql): more suitable default for lql run (hazedav)([73a7132](https://github.com/lacework/go-sdk/commit/73a7132765125f0d2370aef36fa4c8039c1c0cdd))
+* feat(cli/cmd): CLI survey code for azure - ALLY-394 (#720) (djmctavish)([73f8b0c](https://github.com/lacework/go-sdk/commit/73f8b0c83eaf6c67728f871cc3f4e4544c851ba7))
+## Bug Fixes
+* fix: fix fail_on_severity when no vulns are found (#750) (Darren)([479869f](https://github.com/lacework/go-sdk/commit/479869f95ae1c6d15b9108a59fcf36d77d448c4b))
+## Other Changes
+* chore: deprecate CLI Wiki (#739) (Salim Afiune)([eeb6478](https://github.com/lacework/go-sdk/commit/eeb64783f2224c397e61d5d3fe4fd93523e1fc1f))
+* ci: version bump to v0.29.1-dev (Lacework)([fc74684](https://github.com/lacework/go-sdk/commit/fc74684e5143e7a01158f876d28e8a9773c8d2e6))
+---
+# v0.29.0
+
+## Features
+* feat(api): Implement AwsEksAudit CloudAccount interface (#742) (Ross)([879048c](https://github.com/lacework/go-sdk/commit/879048c11430e4606b270be61329a7e71d3138fb))
+## Other Changes
+* chore: update go version to 1.18 (#735) (Darren)([ec936cb](https://github.com/lacework/go-sdk/commit/ec936cb983f4f8dd69a24479252608468cac6a88))
+* ci: version bump to v0.28.1-dev (Lacework)([6d7faae](https://github.com/lacework/go-sdk/commit/6d7faae6b398fd14b7da3b6fdabbb4e7cd772556))
+---
 # v0.28.0
 
 ## Features